eliminate more longs
for 64bit compatibility.

+++ neo/framework/Compressor.cpp
@@ -1726,7 +1726,9 @@ idCompressor_Arithmetic::GetCurrentCount
 */
 int idCompressor_Arithmetic::GetCurrentCount()
 {
-	return ( unsigned int )( ( ( ( ( long ) code - low ) + 1 ) * scale - 1 ) / ( ( ( long ) high - low ) + 1 ) );
+	// DG: use int instead of long for 64bit compatibility
+	return ( unsigned int )( ( ( ( ( int ) code - low ) + 1 ) * scale - 1 ) / ( ( ( int ) high - low ) + 1 ) );
+	// DG end
 }
 
 /*
@@ -1805,9 +1807,11 @@ idCompressor_Arithmetic::RemoveSymbolFromStream
 */
 void idCompressor_Arithmetic::RemoveSymbolFromStream( acSymbol_t* symbol )
 {
-	long range;
+	// DG: use int instead of long for 64bit compatibility
+	int range;
 	
-	range	= ( long )( high - low ) + 1;
+	range	= ( int )( high - low ) + 1;
+	// DG end
 	high	= low + ( unsigned short )( ( range * symbol->high ) / scale - 1 );
 	low		= low + ( unsigned short )( ( range * symbol->low ) / scale );
